tambah delete_by gudang, barang_masuk, laporan_produksi, permintaan_pengiriman, Penerimaan_pengiriman_product, laporan_penjualan

+++ app/Http/Controllers/WarehouseController.php
@@ -14,15 +14,15 @@ class WarehouseController extends Controller
         $q = Warehouse::query()->with('responsible');
 
         if ($request->filled('name')) {
-            $q->where('name', 'like', '%'.$request->name.'%');
+            $q->where('name', 'like', '%' . $request->name . '%');
         }
 
         if ($request->filled('province')) {
-            $q->where('province', 'like', '%'.$request->province.'%');
+            $q->where('province', 'like', '%' . $request->province . '%');
         }
 
         if ($request->filled('city')) {
-            $q->where('city', 'like', '%'.$request->city.'%');
+            $q->where('city', 'like', '%' . $request->city . '%');
         }
 
         if ($request->filled('responsible_id')) {
@@ -41,7 +41,7 @@ class WarehouseController extends Controller
     {
         $path = public_path('assets/data/provinceAndCity.json');
 
-        if (! File::exists($path)) {
+        if (!File::exists($path)) {
             abort(404, 'File provinceAndCity.json tidak ditemukan');
         }
 
@@ -64,7 +64,7 @@ class WarehouseController extends Controller
     {
         $path = public_path('assets/data/provinceAndCity.json');
 
-        if (! File::exists($path)) {
+        if (!File::exists($path)) {
             return response()->json([]);
         }
 
@@ -73,7 +73,7 @@ class WarehouseController extends Controller
 
         $province = collect($data)->firstWhere('province_id', $provinceId);
 
-        if (! $province) {
+        if (!$province) {
             return response()->json([]);
         }
 
@@ -97,6 +97,7 @@ class WarehouseController extends Controller
             'city' => 'required|string',
             'responsible_id' => 'nullable|exists:users,id',
         ]);
+
         Warehouse::create($validated);
 
         return redirect()
@@ -110,7 +111,7 @@ class WarehouseController extends Controller
 
         $path = public_path('assets/data/provinceAndCity.json');
 
-        if (! File::exists($path)) {
+        if (!File::exists($path)) {
             abort(404, 'File provinceAndCity.json tidak ditemukan');
         }
 
@@ -163,7 +164,16 @@ class WarehouseController extends Controller
 
     public function destroy($id)
     {
+        if (!auth()->check()) {
+            abort(403, 'User belum login');
+        }
+
         $warehouse = Warehouse::findOrFail($id);
+
+        $warehouse->update([
+            'deleted_by' => auth()->id(),
+        ]);
+
         $warehouse->delete();
 
         return redirect()
